feat: initial implementation of first-time config form
Initial implementation of the configuration form shown during first time use of the app. This will present default values based on a pack default - this commit does not include support for changing the values.

Currently only supports Browser entities.
Refactor TimlineContainer to receive config as a prop, rather than using the context (which would have to reload, since this is now a nerdlet)